BUG: fix redeclaration of setVh
