+++ pages/summary.js
@@ -46,16 +46,6 @@ export default function summary() {
 	const [locations, setLocations] = useState([]);
 	const [users, setUsers] = useState([]);
 
-	const scriptText = "\
-		const setVh = () => {\
-			let vh = window.innerHeight * 0.01;\
-			document.documentElement.style.setProperty('--vh', `${vh}px`);\
-		};\
-		window.addEventListener('load', setVh);\
-		window.addEventListener('resize', setVh);";
-	useScriptText(scriptText);
-
-
 	useEffect(() => {
 		async function fetchData() {
 			// Get Summary of results
